+++ src/plugins/azoth/plugins/sarin/toxthread.h
@@ -131,7 +131,7 @@ namespace Sarin
 
 		void LoadFriends ();
 
-		void HandleFriendRequest (const uint8_t*, const uint8_t*, uint16_t);
+		void HandleFriendRequest (const uint8_t*, const uint8_t*, size_t);
 		void HandleNameChange (int32_t, const uint8_t*, uint16_t);
 		void UpdateFriendStatus (int32_t);
 		void HandleTypingChange (int32_t, bool);
